requirements (2).txt

Flask
pytesseract
pillow
requests
mysql-connector-python
flask_mysqldb
certifi==2018.11.29
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.0
numpy==1.15.4
opencv-contrib-python==3.4.4.19
pillow>=6.2.0p
ytesseract==0.2.6
werkzeug==0.15.3
